refactor: move configuration handling to new annif.config module
